requirements.txt
pypdf
google-genai
python-dotenv
scikit-learn
numpy
